Update Kafka topic for CAEN AR51 messages
The ECDC Kafka topic name specification states that raw detector readouts should be published on a stream with topic name `{instrument}_detector_samples`, but `nmx_debug` was hard coded for the `MonitorProducer`.
The correct topic name is already constructed and saved as `EFUSettings.KafkaDebugTopic`. 
This commit changes the `MonitorProducer` for CAEN-based modules to use the correct topic name, and updates the `MonitorSerializer` to use the `EFUSettings.DetectorName` as `Source`.